+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -19,6 +19,7 @@ import org.apache.coyote.http11.httprequest.HttpMethod;
 import org.apache.coyote.http11.httprequest.HttpRequest;
 import org.apache.coyote.http11.httpresponse.HttpResponse;
 import org.apache.coyote.http11.httpresponse.HttpStatusCode;
+import org.apache.coyote.http11.parser.FormUrlEncodedHttpRequestParser;
 import org.apache.coyote.http11.parser.HttpResponseParser;
 import org.apache.coyote.http11.parser.SessionParser;
 import org.apache.coyote.http11.session.Session;
@@ -59,7 +60,7 @@ public class Http11Processor implements Runnable, Processor {
 
     private void handleRequest(final BufferedReader reader, final OutputStream outputStream) throws IOException {
         try {
-            final HttpRequest httpRequest = HttpRequest.of(reader);
+            final HttpRequest httpRequest = HttpRequest.of(new FormUrlEncodedHttpRequestParser(reader));
             HttpResponse response;
 
             if (httpRequest.matches(HttpMethod.GET, "/")) {
