+++ tomcat/src/main/java/org/apache/coyote/http11/httprequest/HttpRequest.java
@@ -1,6 +1,5 @@
 package org.apache.coyote.http11.httprequest;
 
-import java.io.BufferedReader;
 import java.io.IOException;
 import org.apache.coyote.http11.cookie.HttpCookie;
 import org.apache.coyote.http11.parser.HttpRequestParser;
@@ -11,12 +10,10 @@ public class HttpRequest {
     private final RequestHeaders requestHeaders;
     private final RequestBody httpRequestBody;
 
-    public static HttpRequest of(final BufferedReader reader) throws IOException {
-        final HttpRequestParser httpRequestParser = new HttpRequestParser(reader);
-
-        final RequestLine requestLine = httpRequestParser.parseRequestLine();
-        final RequestHeaders requestHeaders = httpRequestParser.parseRequestHeaders();
-        final RequestBody requestBody = httpRequestParser.parseRequestBody(requestHeaders);
+    public static HttpRequest of(final HttpRequestParser parser) throws IOException {
+        final RequestLine requestLine = parser.parseRequestLine();
+        final RequestHeaders requestHeaders = parser.parseRequestHeaders();
+        final RequestBody requestBody = parser.parseRequestBody(requestHeaders);
 
         return new HttpRequest(requestLine, requestHeaders, requestBody);
     }
